+++ Sandswept/Enemies/CannonballJellyfish/Skills/SkillDefs/JellyDash.cs
@@ -1,12 +1,13 @@
 using System;
 using Sandswept.Survivors;
 
-namespace Sandswept.Enemies.CannonJellyfish {
+namespace Sandswept.Enemies.CannonballJellyfish.SkillDefs
+{
     public class JellyDash : SkillBase<JellyDash>
     {
         public override string Name => "JellyDash";
 
-        public override string Description => "you shouldnt see this.";
+        public override string Description => "omg hiii :3 :3 :3";
 
         public override Type ActivationStateType => typeof(States.JellyDash);