+++ Sandswept/Enemies/CannonballJellyfish/Skills/States/JellyDash.cs
@@ -3,15 +3,20 @@ using R2API.Utils;
 using RoR2.CharacterAI;
 using System.Linq;
 
-namespace Sandswept.Enemies.CannonJellyfish.States {
+namespace Sandswept.Enemies.CannonballJellyfish.States
+{
     [ConfigSection("Enemies :: Cannonball Jellyfish")]
-    public class JellyDash : BaseState {
+    public class JellyDash : BaseState
+    {
         [ConfigField("Primary - Damage Coefficient", "The damage multiplier to deal.", 6f)]
         public static float DamageCoefficient;
+
         [ConfigField("Primary - Dash Force", "The amount of force to use when dashing", 4000f)]
         public static float DashForce;
+
         //
         private float maxStallDur = 1f;
+
         private OverlapAttack attack;
         private float duration = 0.5f;
         private bool dashedAlready = false;
@@ -22,7 +27,8 @@ namespace Sandswept.Enemies.CannonJellyfish.States {
             base.OnEnter();
             duration += maxStallDur;
 
-            if (base.characterBody.master) {
+            if (base.characterBody.master)
+            {
                 ai = base.characterBody.master.GetComponent<BaseAI>();
             }
 
@@ -35,14 +41,16 @@ namespace Sandswept.Enemies.CannonJellyfish.States {
             attack.teamIndex = base.GetTeam();
             attack.attackerFiltering = AttackerFiltering.NeverHitSelf;
 
-            if (NetworkServer.active) {
+            if (NetworkServer.active)
+            {
                 characterBody.AddBuff(RoR2Content.Buffs.Immune);
             }
 
             FlipComponents();
         }
 
-        public void SetDir() {
+        public void SetDir()
+        {
             base.characterBody.SetAimTimer(0.02f);
         }
 
@@ -52,7 +60,8 @@ namespace Sandswept.Enemies.CannonJellyfish.States {
 
             SetDir();
 
-            if (duration >= maxStallDur && !dashedAlready && base.isAuthority) {
+            if (duration >= maxStallDur && !dashedAlready && base.isAuthority)
+            {
                 dashedAlready = true;
 
                 PhysForceInfo info = new();
@@ -62,16 +71,19 @@ namespace Sandswept.Enemies.CannonJellyfish.States {
                 base.rigidbodyMotor.ApplyForceImpulse(in info);
             }
 
-            if (dashedAlready && base.isAuthority) {
+            if (dashedAlready && base.isAuthority)
+            {
                 attack.Fire();
             }
 
-            if (base.fixedAge >= duration) {
+            if (base.fixedAge >= duration)
+            {
                 outer.SetNextStateToMain();
             }
         }
 
-        public void FlipComponents() {
+        public void FlipComponents()
+        {
             VectorPID[] vPids = base.gameObject.GetComponents<VectorPID>();
             QuaternionPID[] qPids = base.gameObject.GetComponents<QuaternionPID>();
 
@@ -86,7 +98,8 @@ namespace Sandswept.Enemies.CannonJellyfish.States {
             base.OnExit();
             FlipComponents();
             base.rigidbodyMotor.rootMotion = Vector3.zero;
-            if (NetworkServer.active) {
+            if (NetworkServer.active)
+            {
                 characterBody.RemoveBuff(RoR2Content.Buffs.Immune);
             }
         }